MDT and PySCF amplitude grids
The two programs more or less agree up to a multiplicative constant, currently. Big disagreements around small values, which is probably fine
